+++ src/PostgREST/Logger.hs
@@ -161,7 +161,7 @@ observationMessage = \case
   AdminStartObs address ->
     "Admin server listening on " <> address
   AdminServerCrashedObs ex ->
-    "FAILURE: Admin server crashed unexpectedly: " <> (showOnSingleLine '\t' . show) ex
+    "Admin server crashed unexpectedly: " <> (showOnSingleLine '\t' . show) ex
   AppStartObs ver ->
     "Starting PostgREST " <> T.decodeUtf8 ver <> "..."
   AppServerAddressObs address ->